feat: improve ts field pricision to millis for mongodb

+++ src/main/java/io/github/thunderz99/cosmos/util/TimestampUtil.java
@@ -9,22 +9,32 @@ import org.apache.commons.collections4.MapUtils;
  */
 public class TimestampUtil {
 
-    public static void processTimestampPrecision(Map<String, Object> map){
+    /**
+     * convert _ts field in map from Double to Long for compatibility of cosmosdb and mongodb
+     *
+     * <p>
+     * cosmosdb uses Long _ts
+     * mongodb uses Double _ts for sort stability
+     * </p>
+     *
+     * @param map
+     */
+    public static void processTimestampPrecision(Map<String, Object> map) {
 
-        if(MapUtils.isEmpty(map)){
+        if (MapUtils.isEmpty(map)) {
             return;
         }
 
         var ts = map.get("_ts");
 
-        if(ts == null){
+        if (ts == null) {
             // do nothing
             return;
         }
 
-        if(ts instanceof Double){
-            //for compatibility, convert 1502246148.123d to 1502246148
-            map.put("_ts", ((Double) ts).intValue());
+        if (ts instanceof Double) {
+            //for compatibility, convert 1714546148.123d to 1714546148L
+            map.put("_ts", ((Double) ts).longValue());
         }
     }
